[IMP] payroll: move basebrowsableobject to own file
