+++ payroll/models/__init__.py
@@ -7,6 +7,7 @@ from . import hr_salary_rule
 from . import hr_salary_rule_category
 from . import hr_rule_input
 from . import hr_contribution_register
+from . import base_browsable
 from . import hr_payslip
 from . import hr_payslip_line
 from . import hr_payslip_input
